move serialize test
